wpt_babel/content-security-policy/script-src/worker-set-timeout.sub.html

<!DOCTYPE html>
<html>
 <head>
  <meta content="script-src 'self' 'unsafe-inline' 'self' 'unsafe-eval'; connect-src 'self';" http-equiv="Content-Security-Policy"/>
  <title>
   worker-set-timeout
  </title>
  <script src="/resources/testharness.js">
  </script>
  <script src="/resources/testharnessreport.js">
  </script>
  <script src="../support/testharness-helper.js">
  </script>
 </head>
 <body>
  <script>
   "use strict";

assert_worker_is_loaded("./support/worker-with-script-src-none-set-timeout.js", "Dedicated worker delivers its own CSP", "setTimeout blocked");
assert_shared_worker_is_loaded("./support/worker-with-script-src-none-set-timeout.js", "Shared worker delivers its own CSP", "setTimeout blocked");
  </script>
 </body>
</html>
